// packages/provider/src/relayer/gen/chaind.gen.ts
export interface Balance {
  id: number
  contractAddress: string
  contractAlias: string
  accountAddress: string
  tokenID: number
  amountLatest: string
  amountConfirmed: string
  blockHash: string
  blockHeight: number
  
  updatedAt: string
  createdAt: string
}

export interface EtherBalance {
  accountAddress: string
  amountLatestWei: string
  amountConfirmedWei: string
  lastBlockHeight: number
}

export interface GasReceipt {
  gasFee: string
  gasLimit: Array<string>
}

export interface MetaTxn {
  contract: string
  input: string
}

export interface MetaTxnEstimationInput {
  contract: string
  payload: string
  signers: number
}

export interface MetaTxnEstimationOutput {
  payload: string
  gasLimit: number
}

export interface MetaTxnReceipt {
  id: string
  status: string
  revertReason: string
  gasUsed: number
  txnReceipt: string
}

export interface TokenTransfer {
  recipientAddress: string
  tokenID: string
  tokenAmount: string
}

export interface RuntimeStatus {
  chainState: Array<any>
}

export interface SmartWalletContext {
  factory: string
  module: string
  guestModule: string
}

export interface Version {
  webrpcVersion: string
  schemaVersion: string
  schemaHash: string
  appVersion: string
}

export interface Page {
  pageSize?: number
  page?: number
  totalRecords?: number
  column?: string
  before?: any
  after?: any
  sort?: Array<SortBy>
}

export interface SortBy {
  column: string
  order: SortOrder
}

export interface ChaindService {
  ping(headers?: object): Promise<PingReturn>
  version(headers?: object): Promise<VersionReturn>
  runtimeStatus(headers?: object): Promise<RuntimeStatusReturn>
  etherBalance(args: EtherBalanceArgs, headers?: object): Promise<EtherBalanceReturn>
  contractBalanceOf(args: ContractBalanceOfArgs, headers?: object): Promise<ContractBalanceOfReturn>
  contractBalanceOfTokenIDs(args: ContractBalanceOfTokenIDsArgs, headers?: object): Promise<ContractBalanceOfTokenIDsReturn>
  contractBalanceSync(args: ContractBalanceSyncArgs, headers?: object): Promise<ContractBalanceSyncReturn>
  contractBalanceSyncTokenIDs(args: ContractBalanceSyncTokenIDsArgs, headers?: object): Promise<ContractBalanceSyncTokenIDsReturn>
  sendMetaTxn(args: SendMetaTxnArgs, headers?: object): Promise<SendMetaTxnReturn>
  getMetaTxnReceipt(args: GetMetaTxnReceiptArgs, headers?: object): Promise<GetMetaTxnReceiptReturn>
  getMetaTxnNonce(args: GetMetaTxnNonceArgs, headers?: object): Promise<GetMetaTxnNonceReturn>
  getSmartWalletContext(headers?: object): Promise<GetSmartWalletContextReturn>
  getChainID(headers?: object): Promise<GetChainIDReturn>
  tokenFee(headers?: object): Promise<TokenFeeReturn>
  estimateMetaTxnGasReceipt(args: EstimateMetaTxnGasReceiptArgs, headers?: object): Promise<EstimateMetaTxnGasReceiptReturn>
}

export interface PingArgs {
}

export interface PingReturn {
  status: boolean  
}
export interface VersionArgs {
}

export interface VersionReturn {
  version: Version  
}
export interface RuntimeStatusArgs {
}

export interface RuntimeStatusReturn {
  status: RuntimeStatus  
}
export interface EtherBalanceArgs {
  accountAddress: string
}

export interface EtherBalanceReturn {
  balance: EtherBalance  
}
export interface ContractBalanceOfArgs {
  contractHandle: string
  accountAddress: string
}

export interface ContractBalanceOfReturn {
  status: boolean
  balances: Array<Balance>  
}
export interface ContractBalanceOfTokenIDsArgs {
  contractHandle: string
  accountAddress: string
  tokenIDs: Array<number>
}

export interface ContractBalanceOfTokenIDsReturn {
  status: boolean
  balances: Array<Balance>  
}
export interface ContractBalanceSyncArgs {
  contractHandle: string
  accountAddress: string
}

export interface ContractBalanceSyncReturn {
  status: boolean
  balances: Array<Balance>  
}
export interface ContractBalanceSyncTokenIDsArgs {
  contractHandle: string
  accountAddress: string
  tokenIDs: Array<number>
}

export interface ContractBalanceSyncTokenIDsReturn {
  status: boolean
  balances: Array<Balance>  
}
export interface SendMetaTxnArgs {
  call: MetaTxn
}

export interface SendMetaTxnReturn {
  status: boolean
  txnHash: string  
}
export interface GetMetaTxnReceiptArgs {
  metaTxID: string
}

export interface GetMetaTxnReceiptReturn {
  receipt: MetaTxnReceipt  
}
export interface GetMetaTxnNonceArgs {
  walletContractAddress: string
}

export interface GetMetaTxnNonceReturn {
  nonce: string  
}
export interface GetSmartWalletContextArgs {
}

export interface GetSmartWalletContextReturn {
  data: SmartWalletContext  
}
export interface GetChainIDArgs {
}

export interface GetChainIDReturn {
  chainID: number  
}
export interface TokenFeeArgs {
}

export interface TokenFeeReturn {
  isFee: boolean
  fee: string  
}
export interface EstimateMetaTxnGasReceiptArgs {
  call: MetaTxnEstimationInput
  feeToken: string
}

export interface EstimateMetaTxnGasReceiptReturn {
  res: MetaTxnEstimationOutput  
}